Move tool approval test to own file
